requirements.txt
Flask
Flask-SocketIO
GitPython
markdown
markupsafe
openai